Correções - Administration

+++ src/main/java/conselho/estudante/com/projetoconselho/services/administration/CourseService.java
@@ -19,10 +19,13 @@ import conselho.estudante.com.projetoconselho.models.exceptions.DadosDuplicadosE
 import conselho.estudante.com.projetoconselho.models.exceptions.NaoEncontradoException;
 import conselho.estudante.com.projetoconselho.repositories.administration.CourseRepository;
 import conselho.estudante.com.projetoconselho.services.administration.shift.ShiftService;
+import conselho.estudante.com.projetoconselho.services.administration.subject.SubjectService;
 import conselho.estudante.com.projetoconselho.services.logs.CourseLogsService;
 import conselho.estudante.com.projetoconselho.services.users.SupervisorService;
+import conselho.estudante.com.projetoconselho.services.users.TeacherService;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -57,11 +60,25 @@ import java.util.List;
 @AllArgsConstructor
 public class CourseService {
 
+    @Autowired
     private CourseRepository repository;
+    @Autowired
     @Lazy
     private SupervisorService supervisorService;
+    @Autowired
     private CourseLogsService logsService;
+    @Autowired
+    @Lazy
     private ShiftService shiftService;
+    @Autowired
+    @Lazy
+    private TeacherService teacherService;
+    @Autowired
+    @Lazy
+    private SubjectService subjectService;
+    @Autowired
+    @Lazy
+    private ClasseService classeService;
 
     /**
      * Cria um novo curso com os dados fornecidos.
@@ -72,13 +89,12 @@ public class CourseService {
      * @throws DadosDuplicadosException Se um curso com o mesmo nome já existir.
      */
     public CourseResponseDTO create(CourseRequestDTO courseRequestDTO, User actor) {
-        Course course = courseRequestDTO.convert();
+        Course course = courseRequestDTO.convert( shiftService.getObjectShift( courseRequestDTO.shiftId() ) );
         if (repository.existsByName(course.getName())) {
             throw new DadosDuplicadosException("Curso ja cadastrado");
         } else {
             course.setCreatedAt( new Date());
             course = repository.save(course);
-            shiftService.addCourseToShift( course.getShift().getId(), course.getId(), actor);
             logsService.create(actor, course, "create");
             return course.toDTO();
         }
@@ -95,7 +111,7 @@ public class CourseService {
      * @throws NaoEncontradoException Se o curso não for encontrado.
      */
     public CourseResponseDTO update(Long id, CourseRequestDTO courseRequestDTO, User actor) {
-        Course course = courseRequestDTO.convert();
+        Course course = courseRequestDTO.convert( shiftService.getObjectShift( courseRequestDTO.shiftId() ) );
         if (repository.existsById(id)) {
             course.setId(id);
             if (repository.existsByName(course.getName())) {
@@ -140,8 +156,12 @@ public class CourseService {
             changes.add(new ChangeItem("shift", (Object) oldCourse.getShift(), (Object) course.getShift()));
         }
 
-        if ( ! oldCourse.getSupervisor().equals( course.getSupervisor() ) ) {
-            changes.add(new ChangeItem("supervisor", (Object) oldCourse.getSupervisor(), (Object) course.getSupervisor()));
+        if ( oldCourse.getSupervisor() != null ){
+
+            if ( ! oldCourse.getSupervisor().equals( course.getSupervisor() ) ) {
+                changes.add(new ChangeItem("supervisor", (Object) oldCourse.getSupervisor(), (Object) course.getSupervisor()));
+            }
+
         }
 
         return changes;
@@ -227,7 +247,11 @@ public class CourseService {
             throw  new RuntimeException("Supervisor não encontrado");
         }
         Course course = repository.findById(id).get();
-        supervisorService.removeCourse(course.getSupervisor().getId(), course, actor);
+
+        if ( course.getSupervisor() != null ) {
+            supervisorService.removeCourse(course.getSupervisor().getId(), course, actor);
+        }
+
         supervisorService.addCourse(supervisorId, course, actor);
         Supervisor supervisor = supervisorService.findObjectSupervisor(supervisorId);
         logsService.create(actor, course, Collections.singletonList(new ChangeItem("supervisor", (Object) course.getSupervisor(), (Object) supervisor)), "update");
@@ -239,11 +263,12 @@ public class CourseService {
      * Edita o turno de um curso específico.
      *
      * @param id o ID do curso
-     * @param shift o novo turno
+     * @param shiftId o ID do turno
      * @param actor o usuário que editou o curso
      * @return O curso atualizado convertido para DTO.
      */
-    public CourseResponseDTO editShift(Long id, Shift shift, User actor) {
+    public CourseResponseDTO editShift(Long id, Long shiftId, User actor) {
+        Shift shift = shiftService.getObjectShift(shiftId);
         Course course = repository.findById(id).get();
         shiftService.removeCourseOfShift(course.getShift().getId(), course.getId(), actor);
         shiftService.addCourseToShift(shift.getId(), course.getId(), actor);
@@ -299,16 +324,16 @@ public class CourseService {
     /**
      *  Adiciona um professor ao curso.
      *
-     * @param id o ID do curso
-     * @param dto o professor a ser adicionado
+     * @param id o ID do curso ao qual o professor sera adicionado
+     * @param teacherId o ID do professor a ser adicionado
      * @param actor o usuario que adicionou o professor
      */
-    public void addTeacherToCourse(Long id, TeacherRequestDTO dto, User actor) {
+    public void addTeacherToCourse(Long id, Long teacherId, User actor) {
+        Teacher teacher = teacherService.getObjectTeacher(teacherId);
         Course course = repository.findById(id).get();
-        Teacher teacher = dto.convert();
         if(course.addTeacher(teacher)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("teachers", (Object) teacher ) ), "add" );
-            repository.save(course).toDTO();
+            repository.save(course);
         } else {
             throw new NaoEncontradoException("Professor não encontrado");
         }
@@ -324,7 +349,7 @@ public class CourseService {
     public void addTeacherToCourse(Course course, Teacher teacher, User actor) {
         if(course.addTeacher(teacher)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("teachers", (Object) teacher ) ), "add" );
-            repository.save(course).toDTO();
+            repository.save(course);
         } else {
             throw new NaoEncontradoException("Professor não encontrado");
         }
@@ -334,13 +359,13 @@ public class CourseService {
      * Adiciona uma matéria ao curso.
      *
      * @param id O ID do curso ao qual a matéria sera adicionada.
-     * @param dto A matéria a ser adicionada ao curso.
+     * @param subjectId O ID da materia a ser adicionada.
      * @param actor O usuário que adicionou a matéria ao curso.
      * @throws NaoEncontradoException Se a matéria não for encontrada ou não puder ser adicionada.
      */
-    public void addSubjectToCourse(Long id, SubjectRequestDTO dto, User actor) {
+    public void addSubjectToCourse(Long id, Long subjectId, User actor) {
         Course course = repository.findById(id).get();
-        Subject subject = dto.convert();
+        Subject subject = subjectService.getObjectSubject(subjectId);
         if(course.addSubject(subject)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("subjects", (Object) subject ) ), "add" );
             repository.save(course).toDTO();
@@ -372,15 +397,15 @@ public class CourseService {
      * Adiciona uma classe ao curso.
      *
      * @param id o ID do curso ao qual a classe sera adicionada
-     * @param dto a classe a ser adicionada
+     * @param classeId o ID da classe a ser adicionada
      * @param actor o usuario que adicionou a classe
      * @throws NaoEncontradoException Se a classe nao for encontrada ou nao puder ser adicionada.
      * @author Gustavo Stinghen
      * @since 25/03/2024
      */
-    public void addClassToCourse(Long id, ClasseRequestDTO dto, User actor) {
+    public void addClassToCourse(Long id, Long classeId, User actor) {
         Course course = repository.findById(id).get();
-        Classe classe = dto.convert();
+        Classe classe = classeService.findObjectClasse(classeId);
         if(course.addClasse(classe)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("classes", (Object) classe ) ), "add" );
             repository.save(course).toDTO();
@@ -410,13 +435,13 @@ public class CourseService {
      * Remove um professor ao curso.
      *
      * @param id      o ID do curso ao qual o professor sera removido
-     * @param dto     o professor a ser removido
+     * @param teacherId o ID do professor a ser removido
      * @param actor   o usuario que removeu o professor
      * @throws NaoEncontradoException Se a matéria não for encontrada ou não puder ser adicionada.
      */
-    public void removeTeacherFromCourse(Long id, TeacherRequestDTO dto, User actor) {
+    public void removeTeacherFromCourse(Long id, Long teacherId, User actor) {
         Course course = repository.findById(id).get();
-        Teacher teacher = dto.convert();
+        Teacher teacher = teacherService.getObjectTeacher(teacherId);
         if(course.removeTeacher(teacher)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("teachers", (Object) teacher ) ), "remove" );
             repository.save(course).toDTO();
@@ -429,13 +454,13 @@ public class CourseService {
      * Remove uma materia ao curso.
      *
      * @param id o ID do curso ao qual a materia sera removida
-     * @param dto a materia a ser removida
+     * @param subjectId o ID da materia a ser removida
      * @param actor o usuario que removeu a materia
      * @throws NaoEncontradoException Se a matéria nao for encontrada ou nao puder ser adicionada.
      */
-    public void removeSubjectFromCourse(Long id, SubjectRequestDTO dto, User actor) {
+    public void removeSubjectFromCourse(Long id, Long subjectId, User actor) {
         Course course = repository.findById(id).get();
-        Subject subject = dto.convert();
+        Subject subject = subjectService.getObjectSubject(subjectId);
         if(course.removeSubject(subject)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("subjects", (Object) subject ) ), "remove" );
             repository.save(course).toDTO();
@@ -467,15 +492,15 @@ public class CourseService {
      * Remove uma classe ao curso.
      *
      * @param id o ID do curso ao qual a classe sera removida
-     * @param dto a classe a ser removida
+     * @param classeId o ID da classe a ser removida
      * @param actor o usuario que removeu a classe
      * @throws NaoEncontradoException Se a classe nao for encontrada ou nao puder ser adicionada.
      * @author Gustavo Stinghen
      * @since 25/03/2024
      */
-    public void removeClassFromCourse(Long id, ClasseRequestDTO dto, User actor) {
+    public void removeClassFromCourse(Long id, Long classeId, User actor) {
         Course course = repository.findById(id).get();
-        Classe classe = dto.convert();
+        Classe classe = classeService.findObjectClasse(classeId);
         if(course.removeClasse(classe)){
             logsService.create( actor, course, Collections.singletonList( new AddItem("classes", (Object) classe ) ), "remove" );
             repository.save(course).toDTO();
